DeskClock : Remove onDestroy in AlarmActivity
Remove onDestroy to avoid changing alarm state when relaunching
activity happens.

When in power off alarm mode, if relaunch alarm activity during
alarm firing, onDestroy will be called and alarm's state will
be changed into snooze state. And then the alarm instance for
alarm activity will be null because of that power off alarm
instance is the closet missed alarm and the alarm time is changed
after seting into snooze state which is not a missed alarm.

Alarm activity should be displayed after it is relaunched as
relaunch is the normal case for activity. So remove onDestroy.

CRs-Fixed: 967592

Change-Id: I840064475a4e7a6f7e1caff03016ae460870838c